runtime: removing envoy.restart_features.explicit_wildcard_resource
